Create Strange Counter
There is a strange counter. At the first second, it displays the number 3. Each second, the number displayed by decrements by  1 until it reaches 1. In next second, the timer resets to 2*initial number of prior cycle  and continues counting down. The diagram below shows the counter values for each time t in the first three cycles: